+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Ignore parameters from functions annotated with @IBAction.
 - Classes & structs that conform to SwiftUI's PreviewProvider are now retained.
 - Support @main entry points.
+- Fix unused recursive function detection when using the index store.
 
 ## 1.8.0 (2020-10-3)